+++ ydb/library/benchmarks/queries/tpcds/pg/q98.sql
@@ -6,7 +6,7 @@ select i_item_id
       ,i_class
       ,i_current_price
       ,sum(ss_ext_sales_price) as itemrevenue
-      ,sum(ss_ext_sales_price)*100/sum(sum(ss_ext_sales_price)) over
+      ,sum(ss_ext_sales_price)*100::numeric/sum(sum(ss_ext_sales_price)) over
           (partition by i_class) as revenueratio
 from
 	{{store_sales}}
